requirements.txt
pypdf==4.1.0
docx2txt==0.8
google-generativeai==0.4.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-mock==3.12.0
black==24.2.0
pre-commit==3.6.2
flake8==7.0.0
pytest-cov==7.1.0